BUGFIX: log2GammaAlphaPlus should have been *long* double. Assertion was failing

+++ sbm_state.hpp
@@ -63,7 +63,7 @@ struct Labelling {
 	std :: vector< std :: list<int>:: iterator > its; // an iterator into the relevant part of Cluster :: members
 	int NonEmptyClusters;
 	mutable long double SumOfLog2Facts;
-	std :: vector<double> log2GammaAlphaPlus;
+	std :: vector<long double> log2GammaAlphaPlus;
 	Labelling(const int _N, const long double _alpha);
